+++ src/fractal_czi_converters/dev/task_list.py
@@ -8,13 +8,13 @@ DOCS_LINK = "https://github.com/fractal-analytics-platform/fractal-czi-converter
 
 TASK_LIST = [
     ConverterCompoundTask(
-        name="Convert CZI to OME-Zarr",
-        executable_init="czi_single/convert_czi_single_acq_init_task.py",
+        name="Convert CZI Image to OME-Zarr",
+        executable_init="czi_image/convert_czi_image_init_task.py",
         executable="common/single_image_compute_task.py",
         meta_init={"cpus_per_task": 1, "mem": 4000},
         meta={"cpus_per_task": 1, "mem": 12000},
         category="Conversion",
-        tags=["Zeiss", "Single Image Converter"],
+        tags=["Zeiss", "Image Converter"],
     ),
     ConverterCompoundTask(
         name="Convert CZI Plate to OME-Zarr",