+++ tests/test_czi_image_extended.py
@@ -2,10 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from fractal_czi_converters.common import single_image_compute_task
-from fractal_czi_converters.czi_single.convert_czi_single_acq_init_task import (
-    convert_czi_single_acq_init_task,
-)
+from fractal_czi_converters import convert_czi_image
 
 from .utils import run_converter_test
 
@@ -42,7 +39,7 @@ _SINGLE_IMAGE_DATASETS = [
         for name in _SINGLE_IMAGE_DATASETS
     ],
 )
-def test_czi_single_acq_extended(
+def test_czi_image_extended(
     tmp_path: Path,
     init_task_kwargs: dict,
     snapshot_path: Path,
@@ -51,9 +48,8 @@ def test_czi_single_acq_extended(
 ):
     run_converter_test(
         tmp_path=tmp_path,
-        init_task_fn=convert_czi_single_acq_init_task,
-        compute_task_fn=single_image_compute_task,
-        init_task_kwargs=init_task_kwargs,
+        api_fn=convert_czi_image,
+        api_kwargs=init_task_kwargs,
         snapshot_path=snapshot_path,
         update_snapshots=update_snapshots,
         converter_options=converter_options,